transcoding/requirements.txt
absl-py==0.8.1
